pyproject.toml
[project]
name = "hw6"
version = "0.1.0"
description = ""
authors = [
    {name = "i3oi3ka",email = "i3oi3ka@ukr.net"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
